refactor: use dlt as-is
