+++ src/teamster/libraries/dlt/zendesk/pipeline/helpers/talk_api.py
@@ -1,15 +1,11 @@
-"""Zendesk source helpers"""
-
-from datetime import timedelta
 from enum import Enum
 from typing import Any, Dict, Iterator, Optional, Tuple
 
-from dlt.common import pendulum
 from dlt.common.typing import DictStrStr, TDataItems, TSecretValue
 from dlt.sources.helpers.requests import client
 
-from teamster.libraries.dlt.zendesk import settings
-from teamster.libraries.dlt.zendesk.credentials import (
+from teamster.libraries.dlt.zendesk.pipeline import settings
+from teamster.libraries.dlt.zendesk.pipeline.helpers.credentials import (
     TZendeskCredentials,
     ZendeskCredentialsEmailPass,
     ZendeskCredentialsOAuth,
@@ -120,29 +116,3 @@ class ZendeskAPIClient:
                     get_url = response_json["next_page"]
 
             params = {}
-
-
-def make_date_ranges(
-    start: pendulum.DateTime, end: pendulum.DateTime, step: timedelta
-) -> list[tuple[pendulum.DateTime, pendulum.DateTime]]:
-    """Make tuples of (start, end) date ranges between the given `start` and `end`
-    dates.
-    The last range in the resulting list will be capped to the value of `end` argument
-    so it may be smaller than `step`
-
-    Example usage, create 1 week ranges between January 1st 2023 and today:
-    >>> make_date_ranges(pendulum.DateTime(2023, 1, 1).as_tz('UTC'), pendulum.today(),
-    timedelta(weeks=1))
-    """
-    ranges = []
-
-    while True:
-        end_time = min(start + step, end)
-
-        ranges.append((start, end_time))
-        if end_time == end:
-            break
-
-        start = end_time
-
-    return ranges
